requirements.txt
pytesseract
streamlit
google-genai
python-dotenv
pypdf2
python-docx
Pillow